requirements.txt
setuptools~=75.5.0
-e .
fpdf2~=2.8.2
meshio~=5.3.5
numpy~=1.26.4
opencv-python~=4.10.0.84
pupil_apriltags~=1.0.4.post10
pytest~=8.3.4
pyvista~=0.44.2
pyvistaqt~=0.11.1
Pillow~=11.1.0
PyQt5~=5.15.11
scipy~=1.14.1